Reformat code

+++ src/main/java/h12/io/compress/HuffmanCodingCompressor.java
@@ -1,5 +1,10 @@
 package h12.io.compress;
 
+import h12.io.BitOutputStream;
+import h12.util.Bytes;
+import h12.util.TreeNode;
+import org.tudalgo.algoutils.student.annotation.StudentImplementationRequired;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
@@ -7,12 +12,6 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
-
-import org.tudalgo.algoutils.student.annotation.StudentImplementationRequired;
-
-import h12.io.BitOutputStream;
-import h12.util.Bytes;
-import h12.util.TreeNode;
 
 /**
  * Compressor that uses Huffman coding to compress the input.
@@ -61,7 +60,7 @@ public class HuffmanCodingCompressor implements Compressor {
     public void compress() throws IOException {
         HuffmanCoding huffman = new HuffmanCoding();
         String content = getContent();
-        Map<Character, Double> encodingTable = huffman.buildFrequencyTable(content);
+        Map<Character, Integer> encodingTable = huffman.buildFrequencyTable(content);
         TreeNode<Character> root = huffman.buildTree(encodingTable);
         Map<Character, String> encoded = huffman.buildEncodingTable(root);
